fix: use enrforkid in dns
