+++ crates/net/network/src/discovery.rs
@@ -7,12 +7,12 @@ use crate::{
 };
 use enr::Enr;
 use futures::StreamExt;
-use reth_discv4::{DiscoveryUpdate, Discv4, Discv4Config, EnrForkIdEntry};
+use reth_discv4::{DiscoveryUpdate, Discv4, Discv4Config};
 use reth_discv5::{DiscoveredPeer, Discv5};
 use reth_dns_discovery::{
     DnsDiscoveryConfig, DnsDiscoveryHandle, DnsDiscoveryService, DnsNodeRecordUpdate, DnsResolver,
 };
-use reth_primitives::{ForkId, NodeRecord, PeerId};
+use reth_primitives::{EnrForkIdEntry, ForkId, NodeRecord, PeerId};
 use secp256k1::SecretKey;
 use std::{
     collections::VecDeque,
